+++ lib/mastodon_flutter.dart
@@ -4,6 +4,7 @@ export "src/auth_builder.dart";
 export "src/auth_storage.dart";
 export "src/emoji_text_span.dart";
 export 'src/media.dart';
+export 'src/media_info.dart';
 export "src/media_grid.dart";
 export 'src/media_screen.dart';
 export 'src/status_content.dart';